refractor chat related UI and repo
